requirements.txt
Pillow>=9.0.0
keyboard>=0.13.5
pyinstaller>=5.0.0
